Retry/resume strategy added.
