+++ src/main/kotlin/pl/dwojciechowski/service/impl/RemoteServiceImpl.kt
@@ -8,6 +8,7 @@ import io.reactivex.rxjava3.subjects.PublishSubject
 import io.reactivex.rxjava3.subjects.Subject
 import io.rsocket.RSocket
 import pl.dwojciechowski.model.CommandBean
+import pl.dwojciechowski.model.ExecutionStatus
 import pl.dwojciechowski.proto.commands.Command
 import pl.dwojciechowski.proto.commands.CommandServiceClient
 import pl.dwojciechowski.proto.commands.Status
@@ -39,14 +40,14 @@ class RemoteServiceImpl(private val project: Project) : RemoteService {
     override fun executeStreaming(commandBean: CommandBean, doFinally: () -> Unit) {
         try {
             val command = commandBean.getCommand()
-            commandBean.status = CommandBean.ExecutionStatus.RUNNING
+            commandBean.status = ExecutionStatus.RUNNING
             commandBean.response.onNext("Started execution of $commandBean")
             val rSocket = connector.getConnection()
             rSocket.executeStreamingCall(command, commandBean, doFinally)
-            commandBean.actualSubscription = rSocket ?: commandBean.actualSubscription
+            commandBean.actualSubscription = rSocket
             commandSubject.onNext(commandBean)
         } catch (e: Exception) {
-            commandBean.status = CommandBean.ExecutionStatus.STOPPED
+            commandBean.status = ExecutionStatus.STOPPED
             commandBean.response.onNext(e.message)
             doFinally()
             ApplicationManager.getApplication().invokeLater {
@@ -55,20 +56,20 @@ class RemoteServiceImpl(private val project: Project) : RemoteService {
         }
     }
 
-    private fun RSocket?.executeStreamingCall(command: Command, commandBean: CommandBean, doFinally: () -> Unit) {
+    private fun RSocket.executeStreamingCall(command: Command, commandBean: CommandBean, doFinally: () -> Unit) {
         CommandServiceClient(this)
             .executeStreaming(command)
             .doOnNext {
                 commandBean.response.onNext(it.message)
                 if (it.status == Status.FAILED) {
-                    commandBean.status = CommandBean.ExecutionStatus.STOPPED
+                    commandBean.status = ExecutionStatus.STOPPED
                 }
             }.doOnError {
                 commandBean.response.onNext(it.message)
-                commandBean.status = CommandBean.ExecutionStatus.STOPPED
+                commandBean.status = ExecutionStatus.STOPPED
             }.doOnComplete {
-                if (commandBean.status != CommandBean.ExecutionStatus.STOPPED) {
-                    commandBean.status = CommandBean.ExecutionStatus.COMPLETED
+                if (commandBean.status != ExecutionStatus.STOPPED) {
+                    commandBean.status = ExecutionStatus.COMPLETED
                 }
             }.doFinally {
                 doFinally()
